Allow for custom amounts to be donated.
TODO: Fix the charge dollars / cents decision. It'd be nice to have the
client speak cents as the API does but then the POST would have to be
transformed by client side JS to multiply the "Other" amount field by
100.

Added noisebrige logo favicon.

Added inputmode hints to the form fields for mobile devices. This is a
start to mobile optimizing this, eventually w/ flexbox & media-query
stuff we can make it **much** nicer