Beginning of tabbed donation form w/ tab per payment method
Split out the "charge_form" partial into a donation form which will
eventually list the following payment options
  * Credit card
  * Bitcoin
  * PayPal
  * Cash / Cheque (link to Wiki)

Remove ability to set dues flag when creating Subscriptions

Rather than complicate the form w/ this we can instead leave this for
administrative purposes only. We'll notify when a new subscription is
created and we can then ask if people want to account it as dues or not

In the future this will be a timestamp instead of a boolean to allow for
proper allocation as to what charges on a subscriptiotn were dues if
any, depending on the time at which it was changed.