Install devise on the Donor model
Install devise, doing the following
 * Migrate the Donor model to contain the requisite columns
 * Add the devise methods to the Donor class
 * Set the default_url_options for ActiveMailer in dev / prod

Fixed Gemfile source block warning

Add password support for Donor model

Generates random passwords when donors create donations, and sets them
on the Donor.

Plan is to reset the password when users need to sign in in the future,
such as when they want to adjust their subscription or update their
credit card.