Fix updating of Stripe cards
Due to lack of memoization we never set the default card on the customer object
successfully. This issue manifested itself as repeatedly failing Subscription
creation attempts for donors who'd had previous Subscriptions break due to card
expiration. The card creation call worked, but does not set that card to be the
default source. Now we both create and update the default payment source for
the customer.