Render project page templates where appropriate
Allow the cretation of project specific partials in the projects/ views
directory which will be rendered alongside the donation form when the
project page is viewed.

The first page to have such a template will be the laser page.

Add SuggestedDonationAmount class which is used to populate the amounts
drop-down based on the current project being rendered.