Stop using Google fonts; Remove progress bar from homepage
Stop using Google Fonts to import the "Bitter" typeface which we use. Instead we
now vendor it into the application assets.

Remove the progress bar from the homepage as it no longer serves any meaningful
purpose. In the future it'll hopefully be replaced by some monthly statistics to
encourage people to donate further to help us meet our monthly expenses.