Change example name from Edward Snowden
This seems like a small thing, but as much as I like this the way it is,
Snowden is _so_ polarizing, I'd rather not have that debate right as
someone is about to donate to us. Intelligent, caring, thoughtful people
can have wildly opposing opinions on this subject, and having his name
there seems like unnecessarily politicizing the donation process, which
can only hurt it.

Instead, I added Hackers reference.